+++ app/Http/Controllers/VendorController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers {
 
     use Illuminate\Support;
     use App\Http\Requests;
+    use Illuminate\Support\Facades\Gate;
     use Inertia\Response;
     use Illuminate\Http;
     use App\Models;
@@ -17,9 +18,7 @@ namespace App\Http\Controllers {
          */
         public function index(): Response
         {
-            if(auth()->user()->cannot('view_vendors')) {
-                abort(403);
-            }
+            Gate::authorize('view_vendors');
 
             $vendors = Models\Vendor::withoutTrashed()->paginate(10);
             return inertia('vendors/index', compact('vendors'));
@@ -34,15 +33,15 @@ namespace App\Http\Controllers {
          */
         public function show(Models\Vendor $vendor): Response
         {
-            if(auth()->user()->cannot('view_vendor')) {
-                abort(403);
-            }
+            Gate::authorize('view_vendor');
 
             return inertia('vendors/show', compact('vendor'));
         }
 
         public function store(Requests\Vendors\StoreVendor $request): Http\RedirectResponse
         {
+            Gate::authorize('create_vendor');
+
             Models\Vendor::create($request->validated());
 
             return back();
@@ -57,7 +56,10 @@ namespace App\Http\Controllers {
          */
         public function destroy(Models\Vendor $vendor): Http\RedirectResponse
         {
+            Gate::authorize('delete_vendor');
+
             $vendor->delete();
+
             return back();
         }
     }
